langchain_agent/requirements.txt
# LangChain and MCP Adapters
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-mcp-adapters==0.1.9

# Environment and configuration
python-dotenv==1.1.1

# OpenAI (for LLM)
openai==1.98.0

# MCP (Model Context Protocol)
mcp>=1.9.2

# AWS Wrangler for simplified database operations
awswrangler[postgres]>=3.0.0

# Additional dependencies
pydantic>=2.7.4
requests>=2.0.0
aiohttp>=3.8.3
httpx>=0.23.0 